@@ -1,5 +1,7 @@
 {
     "private": true,
+    "name": "vue-3",
+    "description": null,
     "scripts": {
         "dev": "npm run development",
         "development": "mix",
@@ -10,18 +12,30 @@
         "production": "mix --production"
     },
     "devDependencies": {
+        "@vue/compiler-sfc": "^3.2.23",
         "axios": "^0.21",
-        "bootstrap": "^4.6.0",
+        "bootstrap": "^4.6.1",
+        "css-loader": "^6.5.1",
+        "file-loader": "^6.2.0",
         "jquery": "^3.6",
         "laravel-mix": "^6.0.6",
         "lodash": "^4.17.19",
+        "mini-css-extract-plugin": "^2.4.5",
         "popper.js": "^1.16.1",
         "postcss": "^8.1.14",
         "resolve-url-loader": "^3.1.2",
         "sass": "^1.32.11",
         "sass-loader": "^11.0.1",
-        "vue": "^2.6.12",
-        "vue-loader": "^15.9.8",
-        "vue-template-compiler": "^2.6.12"
+        "url-loader": "^4.1.1",
+        "vue": "^3.2.23",
+        "vue-loader": "^16.8.3",
+        "vue-template-compiler": "^2.6.12",
+        "webpack": "^5.64.4",
+        "webpack-cli": "^4.9.1",
+        "webpack-dev-server": "^4.6.0"
+    },
+    "dependencies": {
+        "@popperjs/core": "^2.11.0",
+        "vue-router": "^4.0.12"
     }
 }
