@@ -1,15 +1,22 @@
 <template>
-    <div>
+  <div class="container">
+        <div class="row justify-content-center">
+<!-- page content -->
+      <slot/>
+<!-- page content -->
 
+        </div>
     </div>
 </template>
 
 <script>
     export default {
-        
+
     }
 </script>
 
 <style lang="scss" scoped>
-
+.container {
+background-color: red;
+}
 </style>
